Move mypyc.ops_list to mypyc.primitives.list_ops
