+++ mypyc/irbuild/builder.py
@@ -41,7 +41,7 @@ from mypyc.ops import (
     is_object_rprimitive, FuncSignature, NAMESPACE_MODULE, RaiseStandardError, FuncDecl
 )
 from mypyc.ops_primitive import func_ops
-from mypyc.ops_list import list_append_op, list_len_op, new_list_op, to_list, list_pop_last
+from mypyc.primitives.list_ops import list_append_op, list_len_op, new_list_op, to_list, list_pop_last
 from mypyc.primitives.dict_ops import dict_get_item_op, dict_set_item_op
 from mypyc.primitives.misc_ops import (
     true_op, false_op, iter_op, next_op, py_setattr_op, import_op, get_module_dict_op
