+++ mypyc/irbuild/for_helpers.py
@@ -14,7 +14,7 @@ from mypyc.ops import (
     PrimitiveOp, Branch, Register, AssignmentTarget
 )
 from mypyc.primitives.int_ops import unsafe_short_add
-from mypyc.ops_list import list_len_op, list_get_item_unsafe_op
+from mypyc.primitives.list_ops import list_len_op, list_get_item_unsafe_op
 from mypyc.primitives.misc_ops import iter_op, next_op
 from mypyc.primitives.exc_ops import no_err_occurred_op
