+++ mypyc/irbuild/ll_builder.py
@@ -29,7 +29,7 @@ from mypyc.common import (
     FAST_ISINSTANCE_MAX_SUBCLASSES, MAX_LITERAL_SHORT_INT,
 )
 from mypyc.ops_primitive import binary_ops, unary_ops, method_ops
-from mypyc.ops_list import (
+from mypyc.primitives.list_ops import (
     list_extend_op, list_len_op, new_list_op
 )
 from mypyc.primitives.tuple_ops import list_tuple_op, new_tuple_op
