+++ mypyc/ops_primitive.py
@@ -219,7 +219,7 @@ def custom_op(arg_types: List[RType],
 # Import various modules that set up global state.
 import mypyc.primitives.int_ops  # noqa
 import mypyc.ops_str  # noqa
-import mypyc.ops_list  # noqa
+import mypyc.primitives.list_ops  # noqa
 import mypyc.primitives.dict_ops  # noqa
 import mypyc.primitives.tuple_ops  # noqa
 import mypyc.primitives.misc_ops  # noqa
